Upgrade Selenium to 3.13.0
This brings the major change that the DefaultDriverFactory now
actually creates the default drivers for things. So we do not need to
register it ourselves anymore.

We also use browser specific option objects instead of the factories
on DesiredCapabilities (log messages told us to do so).

Lastly, note that this Selenium upgrade brings binary incompatible
changes. Therefore, we have bumped the version in the previous
commit.